Ozan
Added a name and some dialogue for Ozan (as this is who the character model said he was) based on some of his previous adventures in RS3. I'm unsure whether the code for the pause I've included works or if it'll need tweaking. / Also some minor other dialogue tweaks I noticed while I was there.