Varrock Sewers
Reworking Varrock Sewers to (hopefully..!) add battles for all NPCs, and add dialogue to a handful of others. This is my first attempt at coding in any battles so please let me know if they work. If they work, the levels will need to be adjusted for scaling as I mostly used the same ones for everything.