TzHaar Fight Caves
Added trainer battles for all NPCs within the Fight Caves to simulate the gist of the waves of the main game, along with coding a possible Jad battle as the Gym Leader fight. Includes additional dialogue with NPCs in the TzHaar area to make the area feel more full. Some further additions to make to other roaming NPCs depending on whether these need dialogue or battle encounters.