Melzar's Maze
Coding in NPC dialogue, battles, and items for 1F, 2F, and 3F of the Maze.

Reworking the map for B1F so that it fit in with the redesigned tileset, and then rotating the design 90 degrees to fit in with the upstairs design.

Includes some minor dialogue fixes to places in other parts of the game as flagged in the bug reports section of the Discord.